Move all ssv encoding, decoding and root methods to a separate json_testutils file
