+++ ssv/runner_state.go
@@ -1,9 +1,6 @@
 package ssv
 
 import (
-	"crypto/sha256"
-	"encoding/json"
-
 	"github.com/attestantio/go-eth2-client/spec/phase0"
 
 	"github.com/pkg/errors"
@@ -42,95 +39,4 @@ func (pcs *State) ReconstructBeaconSig(container *PartialSigContainer, root [32]
 		return nil, errors.Wrap(err, "could not reconstruct beacon sig")
 	}
 	return signature, nil
-}
-
-// GetRoot returns the root used for signing and verification
-func (pcs *State) GetRoot() ([32]byte, error) {
-	marshaledRoot, err := pcs.Encode()
-	if err != nil {
-		return [32]byte{}, errors.Wrap(err, "could not encode State")
-	}
-	ret := sha256.Sum256(marshaledRoot)
-	return ret, nil
-}
-
-// Encode returns the encoded struct in bytes or error
-func (pcs *State) Encode() ([]byte, error) {
-	return json.Marshal(pcs)
-}
-
-// Decode returns error if decoding failed
-func (pcs *State) Decode(data []byte) error {
-	return json.Unmarshal(data, &pcs)
-}
-
-func (pcs *State) MarshalJSON() ([]byte, error) {
-
-	// Create alias without duty
-	type StateAlias struct {
-		PreConsensusContainer  *PartialSigContainer
-		PostConsensusContainer *PartialSigContainer
-		RunningInstance        *qbft.Instance
-		DecidedValue           []byte
-		Finished               bool
-		ValidatorDuty          *types.ValidatorDuty `json:"ValidatorDuty,omitempty"`
-		CommitteeDuty          *types.CommitteeDuty `json:"CommitteeDuty,omitempty"`
-	}
-
-	alias := &StateAlias{
-		PreConsensusContainer:  pcs.PreConsensusContainer,
-		PostConsensusContainer: pcs.PostConsensusContainer,
-		RunningInstance:        pcs.RunningInstance,
-		DecidedValue:           pcs.DecidedValue,
-		Finished:               pcs.Finished,
-	}
-
-	if pcs.StartingDuty != nil {
-		if validatorDuty, ok := pcs.StartingDuty.(*types.ValidatorDuty); ok {
-			alias.ValidatorDuty = validatorDuty
-		} else if committeeDuty, ok := pcs.StartingDuty.(*types.CommitteeDuty); ok {
-			alias.CommitteeDuty = committeeDuty
-		} else {
-			return nil, errors.New("can't marshal because BaseRunner.State.StartingDuty isn't ValidatorDuty or CommitteeDuty")
-		}
-	}
-	byts, err := json.Marshal(alias)
-
-	return byts, err
-}
-
-func (pcs *State) UnmarshalJSON(data []byte) error {
-
-	// Create alias without duty
-	type StateAlias struct {
-		PreConsensusContainer  *PartialSigContainer
-		PostConsensusContainer *PartialSigContainer
-		RunningInstance        *qbft.Instance
-		DecidedValue           []byte
-		Finished               bool
-		ValidatorDuty          *types.ValidatorDuty `json:"ValidatorDuty,omitempty"`
-		CommitteeDuty          *types.CommitteeDuty `json:"CommitteeDuty,omitempty"`
-	}
-
-	aux := &StateAlias{}
-
-	// Unmarshal the JSON data into the auxiliary struct
-	if err := json.Unmarshal(data, &aux); err != nil {
-		return err
-	}
-
-	pcs.PreConsensusContainer = aux.PreConsensusContainer
-	pcs.PostConsensusContainer = aux.PostConsensusContainer
-	pcs.RunningInstance = aux.RunningInstance
-	pcs.DecidedValue = aux.DecidedValue
-	pcs.Finished = aux.Finished
-
-	// Determine which type of duty was marshaled
-	if aux.ValidatorDuty != nil {
-		pcs.StartingDuty = aux.ValidatorDuty
-	} else if aux.CommitteeDuty != nil {
-		pcs.StartingDuty = aux.CommitteeDuty
-	}
-
-	return nil
 }
